+++ CHANGELOG.md
@@ -7,6 +7,7 @@ omnibus-gitlab repository.
 
 - Add workhorse metrics to Prometheus
 - Add sidekiq metrics to Prometheus
+- Update default gitlab-shell git timeout to 3 hours ec9ed900
 
 10.2.0