+++ CHANGELOG.md
@@ -9,6 +9,7 @@ omnibus-gitlab repository.
 - Add sidekiq metrics to Prometheus
 - Remove duplicated shared object files from grpc gem (Takuya Noguchi)
 - Update default gitlab-shell git timeout to 3 hours ec9ed900
+- Enable support for git push options in the git config (Romain Maffina) 08edd3e4 
 
 10.2.0